# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpecKit Digital Records API",
        "description": "RESTful API for managing digital records, documents, artifacts, and periodicals",
        "contact": {
            "name": "SpecKit Support",
            "email": "support@speckit.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/periodicals": {
            "get": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "List all periodicals with filters",
                "operationId": "356de26729052c9f1552250d969b3846",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "journal",
                                "magazine",
                                "newspaper",
                                "bulletin"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Create a new periodical",
                "operationId": "4af05a9015454e43a6894553497783f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "organisation_id",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "organisation_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "journal",
                                            "magazine",
                                            "newspaper",
                                            "bulletin"
                                        ]
                                    },
                                    "issn": {
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "type": "string"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "daily",
                                            "weekly",
                                            "monthly",
                                            "quarterly",
                                            "annual"
                                        ]
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Periodical created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/periodicals/{id}": {
            "get": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Get a specific periodical",
                "operationId": "6b681331aa39620db67773e955311946",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Periodical not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Update an existing periodical",
                "operationId": "b1b292337ceeece1884f47bc8bd99da6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "type": "string"
                                    },
                                    "frequency": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Periodical updated successfully"
                    },
                    "404": {
                        "description": "Periodical not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Delete a periodical",
                "operationId": "eb05666bb672b2367901aa42461ded91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Periodical deleted successfully"
                    },
                    "404": {
                        "description": "Periodical not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/periodicals/{id}/issues": {
            "post": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Add an issue to a periodical",
                "operationId": "c663ac483cd702a1d609c8d1ecd5a18e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "issue_number": {
                                        "type": "string"
                                    },
                                    "volume": {
                                        "type": "string"
                                    },
                                    "publication_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "pages": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Issue added successfully"
                    },
                    "404": {
                        "description": "Periodical not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/periodicals/issues/{issueId}/articles": {
            "post": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Add an article to an issue",
                "operationId": "2f4545ace70989bd047725436a9f0744",
                "parameters": [
                    {
                        "name": "issueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "pages": {
                                        "type": "string"
                                    },
                                    "abstract": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article added successfully"
                    },
                    "404": {
                        "description": "Issue not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/periodicals/{id}/subscriptions": {
            "post": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Create a subscription to a periodical",
                "operationId": "de922c6a9ffba91b9f44a05bc49f0012",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "User ID (optional, defaults to authenticated user)",
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "cost": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully"
                    },
                    "404": {
                        "description": "Periodical not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/periodicals/subscriptions/expiring": {
            "get": {
                "tags": [
                    "Periodicals"
                ],
                "summary": "Get subscriptions expiring soon",
                "operationId": "060c69687afe8c85e2c447ff6c3c9bc8",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expiring subscriptions returned"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication token",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}